update dynamics result service to use proper http client factory
